Refactored Code Model
