+++ csharp/inferrer.ts
@@ -1,6 +1,6 @@
 import { processCodeModel } from "#common/process-code-model";
 import { Host } from "@microsoft.azure/autorest-extension-base";
-import { Model } from "remodeler/code-model";
+import { Model } from "#common/code-model/code-model";
 
 export async function process(service: Host) {
   return await processCodeModel(inferStuff, service);
