+++ csharp/lowlevel-generator/generator.ts
@@ -1,6 +1,6 @@
 import { ModelState } from "#common/model-state";
 import { Host, JsonPath } from "@microsoft.azure/autorest-extension-base";
-import { Model } from "remodeler/code-model";
+import { Model } from "#common/code-model/code-model";
 import { Project } from "./project";
 
 export class State extends ModelState<Model> {
