+++ csharp/lowlevel-generator/private-data.ts
@@ -1,7 +1,7 @@
 import { ModelClass } from "#csharp/lowlevel-generator/model/class";
 import { ModelInterface } from "#csharp/lowlevel-generator/model/interface";
 import { PropertyType } from "#csharp/lowlevel-generator/type-declaration";
-import { Dictionary } from "#remodeler/common";
+import { Dictionary } from "#common/dictionary";
 
 export interface CSharpData extends Dictionary<any> {
   typeDeclaration?: PropertyType;
