+++ csharp/lowlevel-generator/validations.ts
@@ -1,6 +1,7 @@
-import { JsonType, Schema } from "#remodeler/code-model";
+import { JsonType} from "#common/code-model/schema";
 import { State } from "./generator";
 import * as message from "./messages";
+import {Schema} from "#common/code-model/schema";
 
 
 export function objectWithFormat(schema: Schema, state: State): boolean {
