+++ powershell/powershell-generator.ts
@@ -1,5 +1,5 @@
 import { deserialize, serialize } from "#common/yaml";
-import { Model } from "#remodeler/code-model";
+import { Model } from "#common/code-model/code-model";
 import { Host } from "@microsoft.azure/autorest-extension-base";
 
 export async function process(service: Host) {
