+++ remodeler/interpretations.ts
@@ -1,7 +1,11 @@
 import * as OpenAPI from "./oai3";
-import { ImplementationLocation, Server, ExternalDocumentation, MediaType, EnumDetails, EnumValue } from "./code-model";
-import { getExtensionProperties, clone, Dictionary } from "./common";
+import { Server} from "../common/code-model/components";
+import { getExtensionProperties, clone} from "./common";
 import { Remodeler } from "./remodeler";
+import {Dictionary} from "#common/dictionary";
+import {EnumDetails, EnumValue} from "#common/code-model/schema";
+import {MediaType} from "#common/code-model/http-operation";
+import {ExternalDocumentation, ImplementationLocation} from "#common/code-model/components";
 
 interface XMSEnum {
   modelAsString?: boolean;
