+++ remodeler/tweak-model-azure.ts
@@ -2,8 +2,9 @@ import { Host, ArtifactMessage, Channel } from "@microsoft.azure/autorest-extens
 import { deserialize, serialize } from "#common/yaml";
 import { processCodeModel } from "#common/process-code-model";
 import { ModelState } from "#common/model-state";
-import { Model, isHttpOperation } from "#remodeler/code-model";
+import { Model } from "#common/code-model/code-model";
 import { map } from "#common/text-manipulation";
+import { isHttpOperation } from "#common/code-model/http-operation";
 
 // Azure version - 
 // Additional tweaks the code model to adjust things so that the code will generate better.
